Reuse existing local WiX installation

+++ cmake/packaging/windows_wix.cmake
@@ -20,15 +20,17 @@ set(WIX_TOOL_PATH "${CMAKE_BINARY_DIR}/.wix")
 file(MAKE_DIRECTORY ${WIX_TOOL_PATH})
 
 # Install WiX locally using dotnet
-execute_process(
-        COMMAND ${DOTNET_EXECUTABLE} tool install --tool-path ${WIX_TOOL_PATH} wix --version ${WIX_VERSION}
-        ERROR_VARIABLE WIX_INSTALL_OUTPUT
-        RESULT_VARIABLE WIX_INSTALL_RESULT
-)
+if(NOT EXISTS "${WIX_TOOL_PATH}/wix${CMAKE_EXECUTABLE_SUFFIX}")
+    execute_process(
+            COMMAND ${DOTNET_EXECUTABLE} tool install --tool-path ${WIX_TOOL_PATH} wix --version ${WIX_VERSION}
+            ERROR_VARIABLE WIX_INSTALL_OUTPUT
+            RESULT_VARIABLE WIX_INSTALL_RESULT
+    )
 
-if(NOT WIX_INSTALL_RESULT EQUAL 0)
-    message(FATAL_ERROR "Failed to install WiX tools locally.
-     WiX packaging may not work correctly, error: ${WIX_INSTALL_OUTPUT}")
+    if(NOT WIX_INSTALL_RESULT EQUAL 0)
+        message(FATAL_ERROR "Failed to install WiX tools locally.
+         WiX packaging may not work correctly, error: ${WIX_INSTALL_OUTPUT}")
+    endif()
 endif()
 
 # Install WiX UI Extension
